requirements.txt
flask==3.0.2
python-dotenv==1.0.1
supabase==2.3.4
streamlit==1.32.0
requests==2.31.0
openai==1.12.0
python-jose==3.3.0
gunicorn==21.2.0